+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- **EXPERIMENTAL** `helper\ParquetDataIterator` provides easy access to iterate over a parquet file (read) and provides full support for nested fields and non-DataFields (Map, List, Struct)
 - **EXPERIMENTAL** `helper\ArrayToDataColumnsConverter` provides 1:1 conversion of PHP (assoc) arrays and a given schema to parquet datacolumns, including nested data - **full support** of Maps, Lists and Structs.
 - **EXPERIMENTAL** `helper\DataColumnsToArrayConverter` provides 1:1 conversion of given parquet datacolumns and a schema to PHP (assoc) arrays, including nested data
 - Compatibility support for UInt64 (partially, limited by PHP)
@@ -9,7 +10,7 @@
 - PHP 8.0 and 8.1 compatibility due to deprecations
 - Better test coverage for some binary readers & writers, unified interfaces
 ### Changed
-- `Field::path` is now an array, stringified field path is now available as `Field::pathString`, set via `Field::setPath(...)` - this is to improve support for field names containing dots and improving handling when using nested and repeated fields
+- `Field::$path` is now an array, stringified field path is now available as `Field::$pathString`, set via `Field::setPath(...)` - this is to improve support for field names containing dots and improving handling when using nested and repeated fields
 - Unsupported compression codec exception message now includes the constant (value) of the codec identifier
 - StructField now fully supports being non-null, nullable or repeated
 - ListField now fully supports being non-null or nullable